Inserido sistema de query builder para comunicacao com o banco de dados

+++ dump.sql
@@ -6,9 +6,16 @@ CREATE TABLE funcionarios (
     sobrenome text,
     cpf bigint NOT NULL UNIQUE,
     setor text NOT NULL,
-    salario_bruto bigint NOT NULL,
+    salario_bruto integer NOT NULL,
     data_admissao date,
     plano_saude boolean DEFAULT true,
     plano_dental boolean DEFAULT true,
     vale_transporte boolean DEFAULT true
 );
+
+-- CREATE TABLE lancamentos (
+--     id serial PRIMARY KEY,
+--     tipo text NOT NULL,
+--     valor integer NOT NULL,
+--     descricao text NOT NULL
+-- );
